+++ src/__tests__/index-test.tsx
@@ -1,5 +1,5 @@
 import {ApolloProvider, ApolloClient, InMemoryCache} from '@apollo/client';
-import {SchemaLink} from '@apollo/link-schema';
+import {SchemaLink} from '@apollo/client/link/schema';
 import {render, waitFor, fireEvent} from '@testing-library/react';
 import React from 'react';
 import 'regenerator-runtime/runtime.js';